feat(ui): add focus management and input handling improvements
- Implement clearFocus and requestFocus methods in UIComponent and UIManager
- Add proper OpenGL state management in renderers
- Update input handling to support double precision coordinates
- Fix layout calculations to handle floating-point dimensions
- Add missing UI event handling methods in InputManager
- Improve text measurement with scaling support
- Enhance lighting system with additional parameters and optimizations

+++ src/com/odyssey/ui/components/Label.java
@@ -172,7 +172,7 @@ public class Label extends UIComponent {
     }
     
     @Override
-    public void setSize(int width, int height) {
+    public void setSize(float width, float height) {
         super.setSize(width, height);
         if (wordWrap && autoSize) {
             updateSize();
